bike-chooser/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
httpx>=0.27.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.30.0
